feat: integrate new services
