+++ billing/src/main/java/cl/perfulandia/billing/BillingApplication.java
@@ -1,0 +1,11 @@
+package cl.perfulandia.billing;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class BillingApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(BillingApplication.class, args);
+    }
+}
