+++ billing/src/main/java/cl/perfulandia/billing/client/SalesClient.java
@@ -1,0 +1,4 @@
+package cl.perfulandia.billing.client;
+
+public class SalesClient {
+}
