+++ billing/src/main/java/cl/perfulandia/billing/controller/InvoiceController.java
@@ -1,0 +1,4 @@
+package cl.perfulandia.billing.controller;
+
+public class InvoiceController {
+}
