+++ billing/src/main/java/cl/perfulandia/billing/dto/SaleDTO.java
@@ -1,0 +1,4 @@
+package cl.perfulandia.billing.dto;
+
+public class SaleDTO {
+}
