+++ billing/src/main/java/cl/perfulandia/billing/model/Invoice.java
@@ -1,0 +1,6 @@
+package cl.perfulandia.billing.model;
+
+
+
+public class Model {
+}
