+++ billing/src/main/java/cl/perfulandia/billing/repository/InvoiceRepository.java
@@ -1,0 +1,4 @@
+package cl.perfulandia.billing.repository;
+
+public class InvoiceRepository {
+}
