+++ billing/src/main/java/cl/perfulandia/billing/service/InvoiceService.java
@@ -1,0 +1,4 @@
+package cl.perfulandia.billing.service;
+
+public class InvoiceService {
+}
