+++ billing/src/test/java/cl/perfulandia/billing/BillingApplicationTests.java
@@ -1,0 +1,11 @@
+package cl.perfulandia.billing;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class BillingApplicationTests {
+    @Test
+    void contextLoads() {
+    }
+}
