--- billing_service/src/main/java/cl/perfulandia/billing/controller/InvoiceController.java
@@ -1,4 +1,0 @@
-package cl.perfulandia.billing.dto;
-
-public class InvoiceController {
-}
