--- billing_service/src/main/java/cl/perfulandia/billing_service/client/SalesClient.java
@@ -1,4 +1,0 @@
-package cl.perfulandia.billing_service.client;
-
-public class SalesClient {
-}
