--- billing_service/src/main/java/cl/perfulandia/billing_service/controller/InvoiceController.java
@@ -1,4 +1,0 @@
-package cl.perfulandia.billing_service.controller;
-
-public class InvoiceController {
-}
