--- billing_service/src/main/java/cl/perfulandia/billing_service/dto/SaleDTO.java
@@ -1,4 +1,0 @@
-package cl.perfulandia.billing_service.dto;
-
-public class SaleDTO {
-}
