--- billing_service/src/main/java/cl/perfulandia/billing_service/model/Invoice.java
@@ -1,6 +1,0 @@
-package cl.perfulandia.billing_service.model;
-
-
-
-public class Model {
-}
