--- billing_service/src/main/java/cl/perfulandia/billing_service/repository/InvoiceRepository.java
@@ -1,4 +1,0 @@
-package cl.perfulandia.billing_service.repository;
-
-public class InvoiceRepository {
-}
