--- billing_service/src/main/java/cl/perfulandia/billing_service/service/InvoiceService.java
@@ -1,4 +1,0 @@
-package cl.perfulandia.billing_service.service;
-
-public class InvoiceService {
-}
