+++ logistics/src/main/java/cl/perfulandia/logistics/LogisticsApplication.java
@@ -1,0 +1,11 @@
+package cl.perfulandia.logistics;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class LogisticsApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(LogisticsApplication.class, args);
+    }
+}
