+++ logistics/src/main/java/cl/perfulandia/logistics/controller/ShipmentController.java
@@ -1,0 +1,4 @@
+package cl.perfulandia.logistics.controller;
+
+public class ShipmentController {
+}
