+++ logistics/src/main/java/cl/perfulandia/logistics/model/Shipment.java
@@ -1,0 +1,4 @@
+package cl.perfulandia.logistics.model;
+
+public class Shipment {
+}
