+++ logistics/src/main/java/cl/perfulandia/logistics/repository/ShipmentRepository.java
@@ -1,0 +1,4 @@
+package cl.perfulandia.logistics.repository;
+
+public class ShipmentRepository {
+}
