+++ logistics/src/main/java/cl/perfulandia/logistics/service/ShipmentService.java
@@ -1,0 +1,4 @@
+package cl.perfulandia.logistics.service;
+
+public class ShipmentService {
+}
