+++ logistics/src/test/java/cl/perfulandia/logistics/LogisticsApplicationTests.java
@@ -1,0 +1,11 @@
+package cl.perfulandia.logistics;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class LogisticsApplicationTests {
+    @Test
+    void contextLoads() {
+    }
+}
