--- logistics_service/src/main/java/cl/perfulandia/logistics_service/controller/ShipmentController.java
@@ -1,4 +1,0 @@
-package cl.perfulandia.logistics_service.controller;
-
-public class ShipmentController {
-}
