--- logistics_service/src/main/java/cl/perfulandia/logistics_service/model/Shipment.java
@@ -1,4 +1,0 @@
-package cl.perfulandia.logistics_service.model;
-
-public class Shipment {
-}
