--- logistics_service/src/main/java/cl/perfulandia/logistics_service/repository/ShipmentRepository.java
@@ -1,4 +1,0 @@
-package cl.perfulandia.logistics_service.repository;
-
-public class ShipmentRepository {
-}
