--- logistics_service/src/main/java/cl/perfulandia/logistics_service/service/ShipmentService.java
@@ -1,4 +1,0 @@
-package cl.perfulandia.logistics_service.service;
-
-public class ShipmentService {
-}
