+++ suppliers/src/main/java/cl/perfulandia/suppliers/SuppliersApplication.java
@@ -1,0 +1,11 @@
+package cl.perfulandia.suppliers;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class SuppliersApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(SuppliersApplication.class, args);
+    }
+}
