+++ suppliers/src/test/java/cl/perfulandia/suppliers/SuppliersApplicationTests.java
@@ -1,0 +1,11 @@
+package cl.perfulandia.suppliers;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class SuppliersApplicationTests {
+    @Test
+    void contextLoads() {
+    }
+}
